+++ Source/Pegasus/Assets Compiler/Fonts/Glyph.cs
@@ -24,7 +24,7 @@
 			Index = index;
 			Character = character;
 			Size = new Size(glyphInfo.bitmap.width, glyphInfo.bitmap.rows);
-			AdvanceX = glyphInfo.advance_x / 64;
+			AdvanceX = (int)glyphInfo.advance_x / 64; // Note: Cast required in Linux 64 bit builds
 			OffsetX = glyphInfo.bitmap_left;
 			OffsetY = glyphInfo.bitmap_top;
